added consent screens

+++ src/utils/installers/tanstack/vite/vite-spa.ts
@@ -6,6 +6,7 @@ import { addViteTSPathAlias } from "@/utils/helpers/config/vite";
 import { removeDirectory } from "@/utils/helpers/fs/directories";
 import Spinnies from "spinnies";
 import { installPackages} from "@/utils/helpers/package-managers";
+import { confirm } from "@inquirer/prompts";
 
 // Define the tailwind schema
 export const tanstackViteReactSchema = z.object({
@@ -23,6 +24,13 @@ export async function addTanstackToVite(bonita_config: TBonitaConfigSchema,) {
   try {
     //  install dependancies
     const config = await promptForTanstackConfig(bonita_config);
+    const consent = await confirm({
+      message: `This will overwrite ${JSON.stringify(bonita_config.vite_tanstack)} Do you want to continue?`,
+      default: true,
+    });
+    if (!consent) {
+      process.exit(1);
+    }
     await setUpRouterTemplate(config);
     await addViteTSPathAlias();
     await removeDirectory("./temp");
